Cut out the function of Parabolic Cylinder Function into another script
