+++ R/RcppExports.R
@@ -1,15 +1,15 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-Int_simt_log_D <- function(x, lamb, a) {
-    .Call(`_neggfl_Int_simt_log_D`, x, lamb, a)
-}
-
 mode_kde <- function(points, T = 512L) {
     .Call(`_neggfl_mode_kde`, points, T)
 }
 
 negfl_bas <- function(y, X, hparams, B) {
     .Call(`_neggfl_negfl_bas`, y, X, hparams, B)
+}
+
+Int_simt_log_D <- function(x, lamb, a) {
+    .Call(`_neggfl_Int_simt_log_D`, x, lamb, a)
 }
