+++ src/RcppExports.cpp
@@ -6,19 +6,6 @@
 
 using namespace Rcpp;
 
-// Int_simt_log_D
-double Int_simt_log_D(double x, double lamb, double a);
-RcppExport SEXP _neggfl_Int_simt_log_D(SEXP xSEXP, SEXP lambSEXP, SEXP aSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< double >::type x(xSEXP);
-    Rcpp::traits::input_parameter< double >::type lamb(lambSEXP);
-    Rcpp::traits::input_parameter< double >::type a(aSEXP);
-    rcpp_result_gen = Rcpp::wrap(Int_simt_log_D(x, lamb, a));
-    return rcpp_result_gen;
-END_RCPP
-}
 // mode_kde
 double mode_kde(arma::vec points, int T);
 RcppExport SEXP _neggfl_mode_kde(SEXP pointsSEXP, SEXP TSEXP) {
@@ -45,11 +32,24 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// Int_simt_log_D
+double Int_simt_log_D(double x, double lamb, double a);
+RcppExport SEXP _neggfl_Int_simt_log_D(SEXP xSEXP, SEXP lambSEXP, SEXP aSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< double >::type x(xSEXP);
+    Rcpp::traits::input_parameter< double >::type lamb(lambSEXP);
+    Rcpp::traits::input_parameter< double >::type a(aSEXP);
+    rcpp_result_gen = Rcpp::wrap(Int_simt_log_D(x, lamb, a));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_neggfl_Int_simt_log_D", (DL_FUNC) &_neggfl_Int_simt_log_D, 3},
     {"_neggfl_mode_kde", (DL_FUNC) &_neggfl_mode_kde, 2},
     {"_neggfl_negfl_bas", (DL_FUNC) &_neggfl_negfl_bas, 4},
+    {"_neggfl_Int_simt_log_D", (DL_FUNC) &_neggfl_Int_simt_log_D, 3},
     {NULL, NULL, 0}
 };
